suchi_phase1_pack/scripts/youtube-transcripts/requirements.txt
youtube-transcript-api>=0.6.0
pyyaml>=6.0
python-dotenv>=1.0.0














